move cdist wrapper into single file
